rewrite join logic

+++ tests/test.sh
@@ -2,8 +2,13 @@
 cd client
 npm run build
 cd ../server
-rm -rf ./node-modules
-npm i
+
+if [ $# -eq 0 ]; then
+  # no args passed
+  rm -rf ./node-modules
+  npm i
+fi
+
 npm run build
 
 npm run test-server 2>&1 ./server.log
